% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{adjust_x_axis_plot}
\alias{adjust_x_axis_plot}
\title{Adjust x-axis}
\usage{
adjust_x_axis_plot(.data, plot)
}
\arguments{
\item{.data}{The raw data.}

\item{plot}{A \code{ggplot2} object.}
}
\description{
Adjusts the x-axis to show rest value at the beginning. Internal use only.
}
\keyword{internal}
