% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_pace}
\alias{convert_to_pace}
\title{Convert m/s to s/100m}
\usage{
convert_to_pace(speed)
}
\arguments{
\item{speed}{The speed to convert to pace.}
}
\value{
the converted pace
}
\description{
It converts speed, in m/s, to pace, in s/100m.
}
\keyword{internal}
