% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo data}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{step}{the number of the step, starting at zero for the baseline.}
\item{length}{the length of each step.}
\item{intensity}{the intensity performed on each step (in this case in watts).}
\item{lactate}{the blood lactate concentration.}
\item{heart_rate}{the heart rate associated with that step.}
}
}
\usage{
demo_data
}
\description{
A dataset containing the lactate and heart rate data collected from a step-incremental
test as an example
}
\keyword{datasets}
