% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helper_data_protocol}
\alias{helper_data_protocol}
\title{Produce data protocol skeleton}
\usage{
helper_data_protocol(
  input_steps,
  input_length_steps,
  input_starting_load,
  input_step_increase,
  input_heart_rate_data,
  input_completed,
  input_last_length_step,
  sport
)
}
\arguments{
\item{input_steps}{The total number of steps from the incremental test.}

\item{input_length_steps}{The length of each step, in minutes.}

\item{input_starting_load}{The starting load.}

\item{input_step_increase}{The step increase.}

\item{input_heart_rate_data}{A boolean indicating whether heart rate data was collected.}

\item{input_completed}{A boolean indicating whether the last step was fully completed.}

\item{input_last_length_step}{If the last step was not fully completed, then indicate how long it lasted.}

\item{sport}{The sport at which the incremental test was performed. One of \code{cycling}, \code{running}, or \code{swimming}.}
}
\description{
This is an internal function used in \code{run_data_input()}.
}
\keyword{internal}
