% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dmax.R
\name{helper_dmax_dmax}
\alias{helper_dmax_dmax}
\title{Helper - Dmax}
\usage{
helper_dmax_dmax(data_prepared, sport, plot)
}
\arguments{
\item{data_prepared}{The data retrieved from \code{prepare_data()}.}

\item{sport}{The sport at which the incremental test was performed. One of \code{cycling}, \code{running}, or \code{swimming}.}

\item{plot}{A boolean to indicate whether to generate a plot from each one of the methods.}
}
\description{
For internal use only and it won't be exported.
}
\keyword{internal}
