% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dmax.R
\name{helper_dmax_log_exp_moddmax}
\alias{helper_dmax_log_exp_moddmax}
\title{Helper - Log-Exp-ModDmax}
\usage{
helper_dmax_log_exp_moddmax(data_prepared, sport, loglog_restrainer, plot)
}
\arguments{
\item{data_prepared}{The data retrieved from \code{prepare_data()}.}

\item{sport}{The sport at which the incremental test was performed. One of \code{cycling}, \code{running}, or \code{swimming}.}

\item{loglog_restrainer}{A scalar from 0 to 1 indicating the percentage of the data that you would like to
restrain for fitting the Log-Log method. For example, \code{1} means no restriction (fits using the whole data), and \code{0.5} means that
only the first 50\% of the data will be used. Default to \code{1}.}

\item{plot}{A boolean to indicate whether to generate a plot from each one of the methods.}
}
\description{
For internal use only and it won't be exported.
}
\keyword{internal}
