% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helper_last_length_step}
\alias{helper_last_length_step}
\title{Calculate length of last step}
\usage{
helper_last_length_step(last_length_step)
}
\arguments{
\item{last_length_step}{A string indicating the length of last step.}
}
\description{
For internal use only.
}
\keyword{internal}
