% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helper_render_table}
\alias{helper_render_table}
\title{Render table for data input}
\usage{
helper_render_table(.data, sport)
}
\arguments{
\item{.data}{The raw data.}

\item{sport}{The sport at which the incremental test was performed. One of \code{cycling}, \code{running}, or \code{swimming}.}
}
\description{
For internal use only.
}
\keyword{internal}
