% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_intensity}
\alias{interpolate_intensity}
\title{Interpolate intensity}
\usage{
interpolate_intensity(.data, interpolation_factor)
}
\arguments{
\item{.data}{The raw data.}

\item{interpolation_factor}{The interpolation factor to be used. This will depend on the sport chosen.}
}
\description{
Interpolate intensity
}
\keyword{internal}
