% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_installed}
\alias{is_installed}
\title{Check if all suggested packages are installed}
\usage{
is_installed()
}
\value{
a boolean
}
\description{
Check if all suggested packages are installed
}
\keyword{internal}
