% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ltp.R
\name{method_ltp}
\alias{method_ltp}
\title{Lactate Turning Point (LTP)}
\usage{
method_ltp(data_prepared, fit, sport, plot)
}
\arguments{
\item{data_prepared}{The data retrieved from \code{prepare_data()}.}

\item{fit}{The fit you would like to use for finding the lactate values associated to each one of the lactate thresholds.}

\item{sport}{The sport at which the incremental test was performed. One of \code{cycling}, \code{running}, or \code{swimming}.}

\item{plot}{A boolean to indicate whether to generate a plot from each one of the methods.}
}
\value{
a \link[tibble:tibble-package]{tibble} with the following columns:
\item{method}{the method used to estimate the lactate threshold}
\item{fitting}{the fitting method used to predict the lactate curve}
\item{intensity}{the intensity associated with the estimated lactate threshold}
\item{lactate}{the lactate concentration associated with the estimated lactate threshold}
\item{heart_rate}{the heart rate associated with the estimated lactate threshold}
\item{plot}{the plot produced to display the lactate threshold}
}
\description{
It applies the \code{LTP} methods: \code{LTP1}, and \code{LTP2}.
}
\keyword{internal}
