% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lactate}
\alias{plot_lactate}
\title{Plot lactate}
\usage{
plot_lactate(data_processed, method)
}
\arguments{
\item{data_processed}{The processed data retrieve within one of the \verb{method_*} functions.}

\item{method}{The lactate threshold method.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plots the lactate method. For internal use only.
}
\keyword{internal}
