% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_methods}
\alias{plot_methods}
\title{Combine lactate threshold methods into one plot}
\usage{
plot_methods(plots, ...)
}
\arguments{
\item{plots}{The \code{ggplot2} objects to be combined.}

\item{...}{Additional arguments passed onto \code{patchwork::wrap_plots()}.}
}
\value{
a \code{patchwork} object
}
\description{
Combine lactate threshold methods into one plot
}
