% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare raw data}
\usage{
prepare_data(.data, intensity_column, lactate_column, heart_rate_column)
}
\arguments{
\item{.data}{The raw data.}

\item{intensity_column}{The name of the intensity column.}

\item{lactate_column}{The name of the lactate column.}

\item{heart_rate_column}{The name of the heart rate column, if applicable.}
}
\value{
a tibble with the following columns:
\item{intensity}{The intensity column.}
\item{lactate}{The lactate column.}
\item{heart_rate}{The heart rate column, if applicable.}
}
\description{
It organizes and renames the raw data.
}
\keyword{internal}
