% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-values.R
\name{retrieve_intensity}
\alias{retrieve_intensity}
\title{Retrieve intensity}
\usage{
retrieve_intensity(data_augmented, fit, model, lactate_value)
}
\arguments{
\item{data_augmented}{The augmented data from the model.}

\item{fit}{The fit you would like to use for finding the lactate values associated to each one of the lactate thresholds.}

\item{model}{The model chosen in \code{fit}.}

\item{lactate_value}{The lactate value.}
}
\value{
the intensity associated with the estimated lactate threshold
}
\description{
Retrieves the intensity associated to a given lactate value.
It is mainly for internal use but it is exported for possible extensions.
}
\keyword{internal}
