% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-input.R
\name{run_data_input}
\alias{run_data_input}
\title{Data input widget}
\usage{
run_data_input(width = 1200, height = 900)
}
\arguments{
\item{width}{The width, in pixels.}

\item{height}{The height, in pixels.}
}
\value{
The code to reproduce the manual data input.
}
\description{
Widget to help with data input.
}
